Fix for Issue #138
After the recent Java SDK upgrade, the getRawMetadataValue() method was added to ObjectMetadata. This method can't be resolved via the simple getter reflection Amazonica uses. The fix is to simply include the s3 namespace from s3transfer, which will result in ObjectMetadata being marshalled without calling that method.

In addition, it looks like a few of the methods on the Download object now require that the download is completed before being invoked, otherwise they throw NPEs.